refactor: Réorganisation des imports et mise à jour des chemins d'accès
- Les chemins d'importation des modules ont été modifiés pour refléter la nouvelle structure du projet, en déplaçant les imports de `app.worker.tasks.base` vers `app.worker.tasks.tasks_utils`.
- Cela améliore la lisibilité et la maintenabilité du code en centralisant les utilitaires de traitement des médias.
- De nouveaux fichiers ont été ajoutés pour les constantes, facilitant ainsi la gestion des valeurs réutilisables dans le code.
