+++ app/globals/others_constants.py
@@ -1,0 +1,5 @@
+class OtherConstants:
+    HLS_PLAYLIST_MASTER_NAME: str = "master.m3u8"
+    HLS_SEGMENTS_DURATION: int = 4      # 4 secondes
+    HLS_DOWNLOAD_FILE_NAME: str = "download.mp4"
+    HLS_PLAYLIST_TYPE: str = "mp4"
