--- app/worker/tasks/base/__init__.py
@@ -1,13 +1,0 @@
-"""Module de base pour le traitement des médias."""
-
-from app.worker.tasks.base.processing_result import ProcessingResult, ProcessingResultType
-from app.worker.tasks.base.processing_step import ProcessingStep
-from app.worker.tasks.base.processing_context import ProcessingContext
-
-__all__ = [
-    "ProcessingResult",
-    "ProcessingResultType",
-    "ProcessingStep",
-    "ProcessingContext",
-]
-
