--- app/worker/tasks/handlers/__init__.py
@@ -1,12 +1,0 @@
-"""Handlers pour le traitement des médias."""
-
-from app.worker.tasks.handlers.progress_handler import ProgressHandler
-from app.worker.tasks.handlers.cleanup_handler import CleanupHandler
-from app.worker.tasks.handlers.error_handler import ErrorHandler
-
-__all__ = [
-    "ProgressHandler",
-    "CleanupHandler",
-    "ErrorHandler",
-]
-
