--- app/worker/tasks/minio/__init__.py
@@ -1,6 +1,0 @@
-"""Module pour les opérations MinIO."""
-
-from app.worker.tasks.minio.minio_operations import MinIOManager
-
-__all__ = ["MinIOManager"]
-
