+++ app/worker/tasks/tasks_utils/base/__init__.py
@@ -1,0 +1,13 @@
+"""Module de base pour le traitement des médias."""
+
+from app.worker.tasks.tasks_utils.base.processing_result import ProcessingResult, ProcessingResultType
+from app.worker.tasks.tasks_utils.base.processing_step import ProcessingStep
+from app.worker.tasks.tasks_utils.base.processing_context import ProcessingContext
+
+__all__ = [
+    "ProcessingResult",
+    "ProcessingResultType",
+    "ProcessingStep",
+    "ProcessingContext",
+]
+
