+++ app/worker/tasks/tasks_utils/base/processing_result.py
