+++ app/worker/tasks/tasks_utils/handlers/__init__.py
@@ -1,0 +1,12 @@
+"""Handlers pour le traitement des médias."""
+
+from app.worker.tasks.tasks_utils.handlers.progress_handler import ProgressHandler
+from app.worker.tasks.tasks_utils.handlers.cleanup_handler import CleanupHandler
+from app.worker.tasks.tasks_utils.handlers.error_handler import ErrorHandler
+
+__all__ = [
+    "ProgressHandler",
+    "CleanupHandler",
+    "ErrorHandler",
+]
+
