+++ app/worker/tasks/tasks_utils/minio/__init__.py
@@ -1,0 +1,6 @@
+"""Module pour les opérations MinIO."""
+
+from app.worker.tasks.tasks_utils.minio.minio_operations import MinIOManager
+
+__all__ = ["MinIOManager"]
+
